requirements.txt
tensorflow>=2.9.0
numpy>=1.21.0
opencv-python>=4.5.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
flask>=2.0.0
flask-socketio>=5.1.0
python-socketio>=5.4.0
python-engineio>=4.2.0
eventlet>=0.33.0
Pillow>=8.3.0
